+++ NEWS.md
@@ -1,9 +1,13 @@
-gosset 1.3 (2024-03-12)
+gosset 1.3 (2024-04-12)
 =========================
 
 ### IMPROVEMENTS
 
 * Implements the function `resample()` to re-sample Plackett-Luce coefficients 
+* Adds standard errors in `regret()`  
+
+###  BUG FIXES
+* Fix the computation of maximized regret in `regret()`. Using `sum(x^2)` instead of `sum(x)^2`
 
 
 gosset 1.2 (2023-07-03)